requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
sqlparse==0.5.5
